Move metricsserializer package to serializer/internal/metrics
