+++ pkg/serializer/internal/metrics/events_test.go
@@ -6,7 +6,7 @@
 //go:build zlib
 // +build zlib
 
-package metricsserializer
+package metrics
 
 import (
 	"encoding/json"
