+++ pkg/serializer/internal/metrics/series.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2022-present Datadog, Inc.
 
-package metricsserializer
+package metrics
 
 import (
 	"bytes"
