+++ pkg/serializer/internal/metrics/series_test.go
@@ -6,7 +6,7 @@
 //go:build zlib
 // +build zlib
 
-package metricsserializer
+package metrics
 
 import (
 	"bytes"
