+++ pkg/serializer/internal/metrics/sketch_benchmark_test.go
@@ -6,7 +6,7 @@
 //go:build test
 // +build test
 
-package metricsserializer
+package metrics
 
 import (
 	"testing"
