+++ pkg/serializer/internal/metrics/test_helper.go
@@ -6,7 +6,7 @@
 //go:build test
 // +build test
 
-package metricsserializer
+package metrics
 
 import (
 	"fmt"
